Added `low_dist.py` and `sort_dist.py`
`low_dist.py` reads kw_trans and returns the distance between all the pairs in `kw_trans`.
CAUTION! The number of all the pairs from n-element set is roughly (n^2)/2. The current `kw_trans` is about 200000+ uniquely lowercased entries. So, this should be used with prior trimming.

`sort_dist.py` reads `rlleven`, the output of `levenshtein.je.py` and returns the sorted such pairs across all japanese head terms.
As default, it sort by the distance between the words, but the flag -n is available to use the 'normalized' distance between the word, which is the distance devided by the larger of the length of the strings.